third_party/blink/web_tests/external/wpt/scroll-animations/css/view-timeline-lookup-expected.txt

This is a testharness.js-based test.
[FAIL] timeline-scope on ancestor sibling, closer timeline wins
  assert_equals: expected "0" but got "auto"
Harness: the test ran to completion.
